Upd: mini refactor (take out adapter from activity, and play with context)
